Updated organization of files
