+++ EFToolkit/Controls/Dialogs/CredentialBox.xaml.cs
@@ -2,6 +2,7 @@ using EFToolkit.Extensions;
 using Microsoft.UI.Xaml.Controls;
 using System;
 using System.Threading.Tasks;
+using EFToolkit.Models;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
