+++ EFToolkit/Controls/Widgets/AcronymBox.xaml.cs
@@ -2,6 +2,7 @@ using Microsoft.UI.Xaml.Controls;
 using System;
 using System.Collections.ObjectModel;
 using System.Linq;
+using EFToolkit.Models;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
