+++ EFToolkit/Controls/Widgets/SchemaBox.xaml.cs
@@ -17,6 +17,7 @@ using System.Threading.Tasks;
 using Windows.ApplicationModel.Contacts.DataProvider;
 using Windows.Foundation;
 using Windows.Foundation.Collections;
+using EFToolkit.Models;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
