+++ EFToolkit/Pages/AcronymPage.xaml.cs
@@ -1,27 +1,16 @@
-using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.WinUI;
-using CommunityToolkit.WinUI.UI.Controls.TextToolbarSymbols;
 using EFToolkit.Controls.Dialogs;
 using EFToolkit.Extensions;
+using EFToolkit.Models;
+using EFToolkit.Enums;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
-using Microsoft.UI.Xaml.Controls.Primitives;
 using Microsoft.UI.Xaml.Data;
 using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
 using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using System.Text.Json;
-using System.Threading.Tasks;
 using Windows.ApplicationModel.DataTransfer;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
-using Windows.Storage;
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
