+++ EFToolkit/Pages/DataVisualizerPage.xaml
@@ -6,7 +6,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:widgets="using:EFToolkit.Controls.Widgets" 
-    xmlns:EFToolkit="using:EFToolkit"
+    xmlns:EFToolkit="using:EFToolkit.Models"
     xmlns:EF="using:EFToolkit.Pages"
     xmlns:controls="using:CommunityToolkit.WinUI.UI.Controls"
     xmlns:controls2="using:CommunityToolkit.WinUI.Controls"
