+++ EFToolkit/Pages/DatabasePage.xaml.cs
@@ -1,10 +1,12 @@
 using EFToolkit.Controls.Dialogs;
 using EFToolkit.Extensions;
+using EFToolkit.Models;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Navigation;
 using System;
 using System.Linq;
+
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
