+++ EFToolkit/Pages/SettingsPage.xaml.cs
@@ -21,6 +21,9 @@ using Windows.System;
 using CommunityToolkit.Mvvm.ComponentModel;
 using System.ComponentModel;
 using EFToolkit.Extensions;
+using EFToolkit.Models;
+using EFToolkit.Enums;
+
 
 // To learn more about WinUI, the WinUI project structure,
 // and more about our project templates, see: http://aka.ms/winui-project-info.
