+++ EFToolkit/Toolkit/Enums/CodeFormatOptions.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EFToolkit.Enums
+{
+    public enum CodeFormatOptions
+    {
+        PascalCase,
+        snake_case,
+    }
+
+}
