+++ EFToolkit/Toolkit/Enums/ModelOptions.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EFToolkit.Enums
+{
+    public enum ModelOptions
+    {
+        Standard,
+        INotifyPropertyChanged,
+        MVVM,
+    }
+
+}
