+++ EFToolkit/Toolkit/Models/SchemaItem.cs
@@ -1,0 +1,17 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EFToolkit.Models
+{
+
+    public partial class SchemaItem : ObservableObject
+    {
+        [ObservableProperty]
+        private string schema;
+    }
+
+}
