+++ EFToolkit/Toolkit/Models/VisualizerItem.cs
@@ -1,0 +1,25 @@
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EFToolkit.Models
+{
+    public partial class VisualizerItem : ObservableObject
+    {
+        [ObservableProperty]
+        private bool? include = true;
+
+        [ObservableProperty]
+        private string columnName = string.Empty;
+
+        [ObservableProperty]
+        private string objectName = string.Empty;
+
+        [ObservableProperty]
+        private string value = string.Empty;
+    }
+
+}
